requirements.txt
bson==0.5.8
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
passlib==1.7.1
pkg-resources==0.0.0
pycrypto==2.6.1
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
